+++ CHANGELOG.md
@@ -40,6 +40,7 @@ UIIN-3437.
 * ECS: Set Held by facet default to current tenant context in Inventory Search. Refs UIIN-2820.
 * Show green message after creating the instance. Fixes UIIN-3446.
 * Fix "Move holdings/items" action item displaying on shared instance. Fixes UIIN-3487.
+* ECS: Set Held by facet default to current tenant context in Inventory Call number Browse. Refs UIIN-3457.
 * Make "data-import-converter-storage" dependency as optional. Refs UIIN-3488.
 
 ## [13.0.10](https://github.com/folio-org/ui-inventory/tree/v13.0.10) (2025-09-01)